@PLN104: satisfy strict clippy (--all-targets --all-features -D warnings)
Four lints in this branch's code:
- control.rs: drop the redundant `as usize` (attributes() already returns usize).
- text.rs text_tl_grow: collapse the nested `if`/`if let` into a let-chain; rewrite the
  upsert `match` as `if let … else`.
- text.rs text_tl_fmt: `#[allow(clippy::option_option)]` — the nested Option encodes three
  real states (off / on-no-fn / on-with-fn), so it is intentional, not collapsible.

fmt + clippy clean; build green; min.loft still leak-free (text.rs behaviour unchanged).

+++ src/parser/control.rs
@@ -8358,7 +8358,7 @@ impl Parser {
                 }
                 if force.contains(d) {
                     let d = *d;
-                    let n_attrs = self.data.attributes(d) as usize;
+                    let n_attrs = self.data.attributes(d);
                     if args.len() < n_attrs {
                         let mut actual = std::mem::take(args);
                         let mut types = vec![Type::Unknown(0); actual.len()];
